Fix type mismatches (#3175)
* Use correct type for arguments to alignTo

alignTo template takes arguments of the same type. Ensure both arguments
passed to it are of actually the same type.

* Use matching types for SFL vector template

* Explicitly return uint32_t instead of size_t for street light style

* Ensure correctly sized type is used when converting offset to pointer

* Ensure correct types when constructing image tables

* Ensure correct types for ObjectManager

* Set ViewportInteraction's target pointer explicitly

+++ src/OpenLoco/src/Objects/ObjectImageTable.cpp
@@ -20,7 +20,7 @@ namespace OpenLoco::ObjectManager
         const auto g1Header = *reinterpret_cast<const Gfx::G1Header*>(remainingData.data());
         remainingData = remainingData.subspan(sizeof(Gfx::G1Header));
 
-        const ImageTableResult res = { _totalNumImages, sizeof(Gfx::G1Header) + g1Header.numEntries * sizeof(Gfx::G1Element32) + g1Header.totalSize };
+        const ImageTableResult res = { _totalNumImages, static_cast<uint32_t>(sizeof(Gfx::G1Header) + g1Header.numEntries * sizeof(Gfx::G1Element32) + g1Header.totalSize) };
 
         if (remainingData.size() < sizeof(Gfx::G1Element32) * g1Header.numEntries)
         {
